requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
duckduckgo-search>=3.9.0
tomli>=2.0.0
pandas>=2.0.0
requests-futures>=1.0.0
colorama>=0.4.6
openpyxl>=3.1.0
holehe>=1.60
phonenumbers>=8.12.0
dnspython>=2.4.0
pyvis>=0.3.0
networkx>=3.0.0
Pillow>=9.0.0
python-whois>=0.9.0
python-multipart>=0.0.6
aiohttp>=3.9.0
